refactor: 💡 organize the export-file-related util, interface,
const code
